Resolved issue #64, added com.xmlcalabash.config.{global,user,local} system properties.
If the global property is set, its value is treated as a filename and
the global configuration will be read from that file instead of from
the configuration stored in the jar file. (This is not recommended.)
The global property value is resolved relative to the current working
directory.

If the user property is set, its value is treated as a filename and
the user configuration will be read from that file. If it is not set,
the value ".calabash" is assumed. The user property value is resolved
relative to the user's home directory. If the user property is the
empty string, no user configuration is loaded.

The local property is interpreted the same way as the user property
except that it is resolved relative to the current working directory.