Work around the new Sys ID behaviour in Saxon 9.7.
In Saxon 9.7 (well, in XSLT 3.0), the System ID of secondary result
trees (using xsl:result-document) are actually the base URI of the
stylesheet (to cut the long story short).  This commit works around that
by wrapping the receiver in one that keeps the initial resolved base URI
of the document.

See http://saxon.markmail.org/thread/33nknzl77x7nn5sl.